fix(guest-access): repair password-policy validation after rebase onto master
The PR was rebased onto a master that had since migrated
ConversationInfo to Jetpack Compose, which broke the build: the merge
resurrected pre-migration files (GuestAccessHelper.kt, its
DialogPasswordBinding/ActivityConversationInfoBinding dialog code) and
dropped an unrelated master fix (@Serializable on
ProvisioningCapability).

- Delete GuestAccessHelper.kt (obsolete View/XML-binding helper no
  longer referenced since the Compose migration) and port its
  password-validation dialog into ConversationInfoActivity's Compose
  screen, replacing the stale XML showPasswordDialog().
- Remove dead, unrelated state left over in ConversationInfoViewModel
  from the old base branch (GetCapabilitiesStartState, CreateRoomUIState)
  and a duplicate ConversationCreationRepository import.
- Restore @Serializable on ProvisioningCapability, lost in the merge.
- Fix ConversationCreationActivity's password-set dialog not dismissing
  after a successful save, and drop an unused ktlint-failing import and
  a misapplied @Suppress in ConversationCreationViewModel.
